requirements.txt
lxml~=4.6.2
PyYAML~=5.4.1
pyperclip~=1.8.2
codeoptimist.yaml~=0.2.0
#pyinstaller~=4.2
